@@ -4,14 +4,14 @@ class CircleColor extends StatelessWidget {
   static const double _kColorElevation = 4.0;
 
   final bool isSelected;
-  final Color color;
-  final VoidCallback onColorChoose;
-  final double circleSize;
-  final double elevation;
-  final IconData iconSelected;
+  final Color? color;
+  final VoidCallback? onColorChoose;
+  final double? circleSize;
+  final double? elevation;
+  final IconData? iconSelected;
 
   const CircleColor({
-    Key key,
+    Key? key,
     @required this.color,
     @required this.circleSize,
     this.onColorChoose,
@@ -20,13 +20,13 @@ class CircleColor extends StatelessWidget {
     this.iconSelected,
   })  : assert(color != null, "You must provide a not null Color"),
         assert(circleSize != null, "CircleColor must have a not null size"),
-        assert(circleSize >= 0, "You must provide a positive size"),
+        assert((circleSize ?? 0) >= 0, "You must provide a positive size"),
         assert(!isSelected || (isSelected && iconSelected != null)),
         super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final brightness = ThemeData.estimateBrightnessForColor(color);
+    final brightness = ThemeData.estimateBrightnessForColor(color!);
     final icon = brightness == Brightness.light ? Colors.black : Colors.white;
 
     return GestureDetector(
@@ -35,7 +35,7 @@ class CircleColor extends StatelessWidget {
         elevation: elevation ?? _kColorElevation,
         shape: const CircleBorder(),
         child: CircleAvatar(
-          radius: circleSize / 2,
+          radius: (circleSize ?? 1) / 2,
           backgroundColor: color,
           child: isSelected ? Icon(iconSelected, color: icon) : null,
         ),
